Fix a bug that caused the background player to stop working
The bug was caused by an incorrect comparison between the existing PlaylistQueue and the new PlaylistQueue. In the handleIntent method in the Player class, both are compared, and if they are considered equal, the existing PlaylistQueue remains unchanged (because why reinitialize the same queue?). However, the overridden equality check was incorrect, as it only checks if the streams are the same, without considering the queueIndex. By inserting this check, the bug is fixed and only if the queues are actually identical, the existing queue is not reinitialized anymore.
I also had to change a test that tests if the equality check disregards the index. However, IMO if you do that the "equal" check doesn't make sense, since these queues cannot be considered equal.

Co-Authored-By: Jared234 <26669009+jared234@users.noreply.github.com>